Potential fix for code scanning alert no. 640: Implicit narrowing conversion in compound assignment (#648)

+++ system/src/main/java/com/percussion/HTTPClient/RespInputStream.java
@@ -59,7 +59,7 @@ final class RespInputStream extends InputStream implements GlobalConstants {
   private int end = 0;
 
   /** the total number of bytes of entity data read from the demux so far */
-  int count = 0;
+  long count = 0L;
 
   static {
     try {
